+++ src/Storage/Storage.Management/Models/PSShare.cs
@@ -39,6 +39,8 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
             this.Etag = share.Etag;
             this.LastModifiedTime = share.LastModifiedTime;
             this.QuotaGiB = share.ShareQuota;
+            this.EnabledProtocols = share.EnabledProtocols;
+            this.RootSquash = share.RootSquash;
             this.Version = share.Version;
             this.Deleted = share.Deleted;
             this.DeletedTime = share.DeletedTime;
@@ -62,6 +64,8 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
             this.Etag = share.Etag;
             this.LastModifiedTime = share.LastModifiedTime;
             this.QuotaGiB = share.ShareQuota;
+            this.EnabledProtocols = share.EnabledProtocols;
+            this.RootSquash = share.RootSquash;
             this.Version = share.Version;
             this.Deleted = share.Deleted;
             this.DeletedTime = share.DeletedTime;
@@ -74,28 +78,26 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
             this.AccessTierStatus = share.AccessTierStatus;
         }
 
-        [Ps1Xml(Label = "ResourceGroupName", Target = ViewControl.List, Position = 0)]
+        [Ps1Xml(Label = "ResourceGroupName", Target = ViewControl.Table, Position = 0)]
         public string ResourceGroupName { get; set; }
 
-        [Ps1Xml(Label = "StorageAccountName", Target = ViewControl.List, Position = 1)]
+        [Ps1Xml(Label = "StorageAccountName", Target = ViewControl.Table, Position = 1)]
         public string StorageAccountName { get; set; }
 
         public string Id { get; set; }
 
-        [Ps1Xml(Label = "Name", Target = ViewControl.List, Position = 2)]
+        [Ps1Xml(Label = "Name", Target = ViewControl.Table, Position = 2)]
         public string Name { get; set; }
 
         public string Type { get; set; }
 
-        [Ps1Xml(Label = "Etag", Target = ViewControl.List, Position = 3)]
         public string Etag { get; set; }
 
-        [Ps1Xml(Label = "QuotaGiB", Target = ViewControl.List, Position = 4)]
+        [Ps1Xml(Label = "QuotaGiB", Target = ViewControl.Table, Position = 3)]
         public int? QuotaGiB { get; set; }
 
         public IDictionary<string, string> Metadata { get; set; }       
 
-        [Ps1Xml(Label = "LastModifiedTime", Target = ViewControl.List, ScriptBlock = "$_.LastModifiedTime.ToString(\"u\")", Position = 5)]
         public DateTime? LastModifiedTime { get; set; }
 
         [Ps1Xml(Label = "Version", Target = ViewControl.List, Position = 7)]